+++ databricks_api/__init__.py
@@ -7,4 +7,3 @@ from databricks_api._version import __license__
 from databricks_api._version import __title__
 from databricks_api._version import __url__
 from databricks_api._version import __version__
-
