+++ databricks_api/_version.py
@@ -2,7 +2,7 @@
 __title__ = "databricks-api"
 __description__ = "Databricks API client interface."
 __url__ = "https://github.com/crflynn/databricks-api"
-__version__ = "0.5.1"
+__version__ = "0.6.0"
 __author__ = "Christopher Flynn"
 __author_email__ = "crf204@gmail.com"
 __license__ = "MIT"
