+++ databricks_api/databricks.py
@@ -1,5 +1,4 @@
 import re
-from pathlib import Path
 
 import databricks_cli.sdk.service as services
 from databricks_cli.sdk import ApiClient
@@ -29,21 +28,3 @@ class DatabricksAPI:
 
         for _, camel_name, service in _get_services():
             setattr(self, camel_name, service(self.client))
-
-
-def _write_stubs():
-    parent_dir = Path(__file__).parent
-    path = Path(parent_dir, "databricks").with_suffix(".pyi")
-    with open(path, "w") as stub_file:
-        stub_file.write("from databricks_cli.sdk import ApiClient\n")
-        stub_file.write("from databricks_cli.sdk.service import *\n\n")
-        stub_file.write(f"class {DatabricksAPI.__name__}:\n")
-        stub_file.write("\tdef __init__(self, **kwargs):\n")
-        stub_file.write("\t\tself.client: ApiClient = ...\n")
-        stub_file.write("\n".join(
-            f"\t\tself.{camel_name}: {service_name} = ..." for service_name, camel_name, service in _get_services()
-        ))
-
-
-if __name__ == "__main__":
-    _write_stubs()
