+++ databricks_api/databricks.pyi
@@ -2,16 +2,16 @@ from databricks_cli.sdk import ApiClient
 from databricks_cli.sdk.service import *
 
 class DatabricksAPI:
-	def __init__(self, **kwargs):
-		self.client: ApiClient = ...
-		self.jobs: JobsService = ...
-		self.cluster: ClusterService = ...
-		self.policy: PolicyService = ...
-		self.managed_library: ManagedLibraryService = ...
-		self.dbfs: DbfsService = ...
-		self.workspace: WorkspaceService = ...
-		self.secret: SecretService = ...
-		self.groups: GroupsService = ...
-		self.token: TokenService = ...
-		self.instance_pool: InstancePoolService = ...
-		self.delta_pipelines: DeltaPipelinesService = ...
+    def __init__(self, **kwargs):
+        self.client: ApiClient = ...
+        self.jobs: JobsService = ...
+        self.cluster: ClusterService = ...
+        self.policy: PolicyService = ...
+        self.managed_library: ManagedLibraryService = ...
+        self.dbfs: DbfsService = ...
+        self.workspace: WorkspaceService = ...
+        self.secret: SecretService = ...
+        self.groups: GroupsService = ...
+        self.token: TokenService = ...
+        self.instance_pool: InstancePoolService = ...
+        self.delta_pipelines: DeltaPipelinesService = ...
